refactor: customer, wish를 같은 도메인으로 판단하여,패키지를 customer로 통합  (ios-h/airbnb#37)
- 시그리드 3번째 리뷰 중에 도메인과 Entity를 구분하여 생각하라는 의견으로 패키지구분에 대해 대답한 부분에 따봉을 눌러주셔서 반영
