+++ BE/src/main/java/org/team4/airbnb/customer/Customer.java
@@ -14,7 +14,6 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
-import org.team4.airbnb.wish.Wish;
 
 @Entity
 @Builder
