+++ BE/src/main/java/org/team4/airbnb/customer/dto/WishResponse.java
@@ -1,4 +1,4 @@
-package org.team4.airbnb.wish.dto;
+package org.team4.airbnb.customer.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
