+++ BE/src/test/java/org/team4/airbnb/customer/CustomerRepositoryTest.java
@@ -15,8 +15,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.team4.airbnb.accommodation.Accommodation;
 import org.team4.airbnb.accommodation.AccommodationRepository;
 import org.team4.airbnb.util.EntityCreator;
-import org.team4.airbnb.wish.Wish;
-import org.team4.airbnb.wish.WishRepository;
 
 @DataJpaTest
 @Transactional
